rename linkregistrationform and registrationaccordion
